requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
discord.py==2.3.2
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt